requirements.txt
torchvision
numpy
torch
scikit-learn
matplotlib
seaborn
tqdm
opencv-python
pyyaml
git+https://github.com/facebookresearch/sam2.git#egg=sam-2
